+++ crates/tribal-worker/src/prompt.rs
@@ -15,6 +15,6 @@ pub(crate) use relation::{
     CandidateOutcome, RelationPromptContext, SimilarItemDecisionContext, assemble_relation_prompt,
     relation_user_context,
 };
-pub use renderer::reserved_keys;
 pub(crate) use triage::{SimilarItemContext, assemble_triage_prompt, triage_user_context};
 pub use validation::synthetic_validation_context;
+pub use variables::reserved_keys;
